NEW - bug 321852: [Class Diagram] Can not drop class in a package https://bugs.eclipse.org/bugs/show_bug.cgi?id=321852

+++ plugins/uml/org.eclipse.papyrus.diagram.clazz/custom-src/org/eclipse/papyrus/diagram/clazz/custom/policies/ClassDiagramDragDropEditPolicy.java
@@ -257,12 +257,14 @@ public class ClassDiagramDragDropEditPolicy extends CommonDiagramDragDropEditPol
 		CompositeCommand cc = new CompositeCommand(CONTAINED_CLASS_DROP_TO_COMPARTMENT);
 		cc = getDefaultDropNodeCommand(nodeVISUALID, dropRequest.getLocation(), droppedElement);
 
-		View droppedView = (View)containmentHelper.findEditPartFor(getViewer().getEditPartRegistry(), droppedElement).getModel();
+		if((View)containmentHelper.findEditPartFor(getViewer().getEditPartRegistry(), droppedElement)!=null){
+			View droppedView = (View)containmentHelper.findEditPartFor(getViewer().getEditPartRegistry(), droppedElement).getModel();
 
-		containmentHelper.deleteIncomingContainmentLinksFor(cc, droppedView);
+			containmentHelper.deleteIncomingContainmentLinksFor(cc, droppedView);
 
-		// Delete the dropped element existing outside the compartment
-		cc.add(new DeleteCommand(getEditingDomain(), droppedView));
+			// Delete the dropped element existing outside the compartment
+			cc.add(new DeleteCommand(getEditingDomain(), droppedView));
+		}
 		return new ICommandProxy(cc);
 	}
